requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
python-decouple==3.8
django-cors-headers==4.3.1
PyJWT==2.8.0
requests==2.31.0
whitenoise==6.6.0